+++ src/Libraries/SmartStore.Data/Migrations/201506261756463_PrimaryStoreCurrencyMultiStore.Designer.cs
@@ -7,13 +7,13 @@ namespace SmartStore.Data.Migrations
     using System.Resources;
     
     [GeneratedCode("EntityFramework.Migrations", "6.1.0-30225")]
-    public sealed partial class QueuedEmailAttachments : IMigrationMetadata
+    public sealed partial class PrimaryStoreCurrencyMultiStore : IMigrationMetadata
     {
-        private readonly ResourceManager Resources = new ResourceManager(typeof(QueuedEmailAttachments));
+        private readonly ResourceManager Resources = new ResourceManager(typeof(PrimaryStoreCurrencyMultiStore));
         
         string IMigrationMetadata.Id
         {
-            get { return "201506262044329_QueuedEmailAttachments"; }
+            get { return "201506261756463_PrimaryStoreCurrencyMultiStore"; }
         }
         
         string IMigrationMetadata.Source