+++ src/Libraries/SmartStore.Data/Setup/SeedData/InvariantSeedData.cs
@@ -73,6 +73,10 @@ namespace SmartStore.Data.Setup
 			var seName = GetSeName("company-logo");
 			var imgCompanyLogo = _ctx.Set<Picture>().Where(x => x.SeoFilename == seName).FirstOrDefault();
 			
+			var currency = _ctx.Set<Currency>().FirstOrDefault(x => x.CurrencyCode == "EUR");
+			if (currency == null)
+				currency = _ctx.Set<Currency>().First();
+			
 			var entities = new List<Store>()
 			{
 				new Store()
@@ -82,7 +86,9 @@ namespace SmartStore.Data.Setup
 					Hosts = "yourstore.com,www.yourstore.com",
 					SslEnabled = false,
 					DisplayOrder = 1,
-					LogoPictureId = imgCompanyLogo.Id
+					LogoPictureId = imgCompanyLogo.Id,
+					PrimaryStoreCurrencyId = currency.Id,
+					PrimaryExchangeRateCurrencyId = currency.Id
 				}
 			};
 			this.Alter(entities);
@@ -4244,8 +4250,6 @@ namespace SmartStore.Data.Setup
 				},
 				new CurrencySettings()
 				{
-					PrimaryStoreCurrencyId = _ctx.Set<Currency>().First().Id,
-					PrimaryExchangeRateCurrencyId = _ctx.Set<Currency>().First().Id,
 				},
 				new MeasureSettings()
 				{