+++ src/Plugins/SmartStore.GoogleMerchantCenter/Services/GoogleFeedService.cs
@@ -14,6 +14,7 @@ using SmartStore.Core.Domain.Directory;
 using SmartStore.Core.Logging;
 using SmartStore.GoogleMerchantCenter.Domain;
 using SmartStore.GoogleMerchantCenter.Models;
+using SmartStore.Services;
 using SmartStore.Services.Catalog;
 using SmartStore.Services.Directory;
 using SmartStore.Services.Localization;
@@ -35,6 +36,8 @@ namespace SmartStore.GoogleMerchantCenter.Services
 		private readonly MeasureSettings _measureSettings;
 		private readonly IDbContext _dbContext;
 		private readonly AdminAreaSettings _adminAreaSettings;
+		private readonly ICurrencyService _currencyService;
+		private readonly ICommonServices _services;
 
 		public GoogleFeedService(
 			IRepository<GoogleProductRecord> gpRepository,
@@ -45,6 +48,8 @@ namespace SmartStore.GoogleMerchantCenter.Services
 			MeasureSettings measureSettings,
 			IDbContext dbContext,
 			AdminAreaSettings adminAreaSettings,
+			ICurrencyService currencyService,
+			ICommonServices services,
 			IComponentContext ctx)
         {
             _gpRepository = gpRepository;
@@ -55,6 +60,8 @@ namespace SmartStore.GoogleMerchantCenter.Services
 			_measureSettings = measureSettings;
 			_dbContext = dbContext;
 			_adminAreaSettings = adminAreaSettings;
+			_currencyService = currencyService;
+			_services = services;
 
 			_helper = new FeedPluginHelper(ctx, "SmartStore.GoogleMerchantCenter", "SmartStore.GoogleMerchantCenter", () =>
 			{
@@ -387,7 +394,7 @@ namespace SmartStore.GoogleMerchantCenter.Services
 				writer.WriteElementString("g", "condition", _googleNamespace, Condition());
 				writer.WriteElementString("g", "availability", _googleNamespace, Availability(product));
 
-				decimal price = Helper.GetProductPrice(product, currency);
+				decimal price = Helper.GetProductPrice(product, currency, fileCreation.Store);
 				string specialPriceDate;
 
 				if (SpecialPrice(product, out specialPriceDate))
@@ -398,7 +405,7 @@ namespace SmartStore.GoogleMerchantCenter.Services
 					// get regular price ignoring any special price
 					decimal specialPrice = product.SpecialPrice.Value;
 					product.SpecialPrice = null;
-					price = Helper.GetProductPrice(product, currency);
+					price = Helper.GetProductPrice(product, currency, fileCreation.Store);
 					product.SpecialPrice = specialPrice;
 
 					_dbContext.SetToUnchanged<Product>(product);
@@ -690,8 +697,11 @@ namespace SmartStore.GoogleMerchantCenter.Services
 						VisibleIndividuallyOnly = true
 					};
 
-					var currency = Helper.GetUsedCurrency(Settings.CurrencyId);
+					var currency = _currencyService.GetCurrencyById(Settings.CurrencyId);
 					var measureWeightSystemKey = _measureService.GetMeasureWeightById(_measureSettings.BaseWeightId).SystemKeyword;
+
+					if (currency == null || !currency.Published)
+						currency = _services.WorkContext.WorkingCurrency;
 
 					writer.WriteStartDocument();
 					writer.WriteStartElement("rss");