+++ src/Plugins/SmartStore.Shipping/Controllers/ByTotalController.cs
@@ -1,13 +1,11 @@
 ﻿using System.Web.Mvc;
 using SmartStore.Core.Domain.Common;
-using SmartStore.Core.Domain.Directory;
+using SmartStore.Services;
+using SmartStore.Services.Directory;
+using SmartStore.Services.Shipping;
 using SmartStore.Shipping.Domain;
 using SmartStore.Shipping.Models;
 using SmartStore.Shipping.Services;
-using SmartStore.Services.Configuration;
-using SmartStore.Services.Directory;
-using SmartStore.Services.Shipping;
-using SmartStore.Services.Stores;
 using SmartStore.Web.Framework.Controllers;
 using Telerik.Web.Mvc;
 
@@ -17,34 +15,26 @@ namespace SmartStore.Shipping.Controllers
     public class ByTotalController : PluginControllerBase
     {
         private readonly IShippingService _shippingService;
-		private readonly IStoreService _storeService;
-        private readonly ISettingService _settingService;
         private readonly IShippingByTotalService _shippingByTotalService;
         private readonly ShippingByTotalSettings _shippingByTotalSettings;
         private readonly ICountryService _countryService;
-        private readonly ICurrencyService _currencyService;
-        private readonly CurrencySettings _currencySettings;
 		private readonly AdminAreaSettings _adminAreaSettings;
+		private readonly ICommonServices _services;
 
-        public ByTotalController(IShippingService shippingService,
-			IStoreService storeService, 
-            ISettingService settingService, 
+        public ByTotalController(
+			IShippingService shippingService,
             IShippingByTotalService shippingByTotalService,
             ShippingByTotalSettings shippingByTotalSettings, 
             ICountryService countryService,
-            ICurrencyService currencyService, 
-            CurrencySettings currencySettings,
-			AdminAreaSettings adminAreaSettings)
+			AdminAreaSettings adminAreaSettings,
+			ICommonServices services)
         {
             this._shippingService = shippingService;
-			this._storeService = storeService;
-            this._settingService = settingService;
             this._shippingByTotalService = shippingByTotalService;
             this._shippingByTotalSettings = shippingByTotalSettings;
             this._countryService = countryService;
-            this._currencyService = currencyService;
-            this._currencySettings = currencySettings;
 			this._adminAreaSettings = adminAreaSettings;
+			this._services = services;
         }
 
         public ActionResult Configure()
@@ -56,29 +46,31 @@ namespace SmartStore.Shipping.Controllers
             }
 
             var model = new ByTotalListModel();
+			var allStores = _services.StoreService.GetAllStores();
+
             foreach (var sm in shippingMethods)
             {
-                model.AvailableShippingMethods.Add(new SelectListItem() { Text = sm.Name, Value = sm.Id.ToString() });
+                model.AvailableShippingMethods.Add(new SelectListItem { Text = sm.Name, Value = sm.Id.ToString() });
             }
 
 			//stores
-			model.AvailableStores.Add(new SelectListItem() { Text = "*", Value = "0" });
-			foreach (var store in _storeService.GetAllStores())
+			model.AvailableStores.Add(new SelectListItem { Text = "*", Value = "0" });
+			foreach (var store in allStores)
 			{
-				model.AvailableStores.Add(new SelectListItem() { Text = store.Name, Value = store.Id.ToString() });
+				model.AvailableStores.Add(new SelectListItem { Text = store.Name, Value = store.Id.ToString() });
 			}
 
-            //model.AvailableCountries.Add(new SelectListItem() { Text = "*", Value = "0" });
+            //model.AvailableCountries.Add(new SelectListItem { Text = "*", Value = "0" });
             var countries = _countryService.GetAllCountries(true);
             foreach (var c in countries)
             {
-                model.AvailableCountries.Add(new SelectListItem() { Text = c.Name, Value = c.Id.ToString() });
+                model.AvailableCountries.Add(new SelectListItem { Text = c.Name, Value = c.Id.ToString() });
             }
 
             model.LimitMethodsToCreated = _shippingByTotalSettings.LimitMethodsToCreated;
             model.SmallQuantityThreshold = _shippingByTotalSettings.SmallQuantityThreshold;
             model.SmallQuantitySurcharge = _shippingByTotalSettings.SmallQuantitySurcharge;
-            model.PrimaryStoreCurrencyCode = _currencyService.GetCurrencyById(_currencySettings.PrimaryStoreCurrencyId).CurrencyCode;
+            model.PrimaryStoreCurrencyCode = _services.StoreContext.CurrentStore.PrimaryStoreCurrency.CurrencyCode;
 			model.GridPageSize = _adminAreaSettings.GridPageSize;
 
             return View(model);
@@ -166,7 +158,7 @@ namespace SmartStore.Shipping.Controllers
             _shippingByTotalSettings.SmallQuantityThreshold = model.SmallQuantityThreshold;
             _shippingByTotalSettings.SmallQuantitySurcharge = model.SmallQuantitySurcharge;
 
-            _settingService.SaveSetting(_shippingByTotalSettings);
+            _services.Settings.SaveSetting(_shippingByTotalSettings);
 
             return Json(new { Result = true });
         }