+++ src/Presentation/SmartStore.Web.Framework/Controllers/RssActionResult.cs
@@ -12,8 +12,10 @@ namespace SmartStore.Web.Framework.Controllers
         {
             context.HttpContext.Response.ContentType = "application/rss+xml";
 
-            var rssFormatter = new Rss20FeedFormatter(Feed);
-            using (var writer = XmlWriter.Create(context.HttpContext.Response.Output))
+            var rssFormatter = new Rss20FeedFormatter(Feed, false);
+			var settings = new XmlWriterSettings { Indent = true, IndentChars = "\t", CheckCharacters = false };
+
+            using (var writer = XmlWriter.Create(context.HttpContext.Response.Output, settings))
             {
                 rssFormatter.WriteTo(writer);
             }