+++ src/Presentation/SmartStore.Web/Administration/Models/Settings/BlogSettingsModel.cs
@@ -20,6 +20,9 @@ namespace SmartStore.Admin.Models.Settings
         [SmartResourceDisplayName("Admin.Configuration.Settings.Blog.NumberOfTags")]
         public int NumberOfTags { get; set; }
 
+		[SmartResourceDisplayName("Admin.Configuration.Settings.Blog.MaxAgeInDays")]
+		public int MaxAgeInDays { get; set; }
+
         [SmartResourceDisplayName("Admin.Configuration.Settings.Blog.ShowHeaderRSSUrl")]
         public bool ShowHeaderRssUrl { get; set; }
     }